lib: pdn: decrease stack and heap for pdn_dynamic_info_get()
Decrease stack requirements for pdn_dynamic_info_get(). Remove use
of k_malloc(). Add use of pdn_dynamic_info_get() to PDN sample and
add tests.

+++ include/modem/pdn.h
@@ -243,7 +243,7 @@ __deprecated int pdn_dynamic_params_get(uint8_t cid, struct in_addr *dns4_pri,
  *
  * @return Zero on success or an error code on failure.
  */
-int pdn_dynamic_info_get(uint32_t cid, struct pdn_dynamic_info *pdn_info);
+int pdn_dynamic_info_get(uint8_t cid, struct pdn_dynamic_info *pdn_info);
 
 /**
  * @brief Retrieve the default Access Point Name (APN).
